+++ browser.js
@@ -68,6 +68,8 @@ export {
   validatePlugin,
   collectPluginModuleCapabilities,
   listPluginModuleCapabilities,
+  collectPluginWorkspaceTemplates,
+  listPluginWorkspaceTemplates,
 } from './plugins/index.js';
 
 import {